+++ MaximumEntropyThreshold/MaximumEntropyThreshold.py
@@ -74,7 +74,7 @@ def main():
         # 레이어 가져오기 // Get image layers
         layer1 = viewImage[0].GetLayer(0)
         layer2 = viewImage[1].GetLayer(0)
-        flpTemp = CFLPoint[float](0, 0)
+        flpTemp = CFLPoint[Double](0, 0)
 
         # Text 출력 // Display text
         if (res := layer1.DrawTextImage(flpTemp, "Source Image", EColor.YELLOW, EColor.BLACK, 20)).IsFail():
